\name{Iran.non.Oil.GDP.Cycle}
\docType{data}
\alias{Iran.non.Oil.GDP.Cycle}
\title{
Cycle of non-Oil GDP of Iran.}
\description{
Cycle of non-Oil GDP of Iran. (Non-Oil GDP after x12, and HP filtering)
}
\usage{
Iran.non.Oil.GDP.Cycle
}
\format{ts Quarterly Time Series}
\source{Central Bank of Islamic Republic of Iran. Further calculations by Majid Einian}
\references{Einian, M. and M. Barakchian (2012), Measuring and Dating Business Cycles of the Economy of Iran, \emph{Pool va Eqtesad}, forthcoming (in Persian)}