\name{Iran.non.Oil.GDP.Quarterly.Growth}
\docType{data}
\alias{Iran.non.Oil.GDP.Quarterly.Growth}
\title{
Quartely Grwoth of non-Oil GDP of Iran.}
\description{
Quartely Grwoth of non-Oil GDP of Iran. (after x12)
}
\usage{
Iran.non.Oil.GDP.Quartely.Grwoth
}
\format{ts Quarterly Time Series}
\source{Central Bank of Islamic Republic of Iran. Further calculations by Majid Einian}
\references{Einian, M. and M. Barakchian (2014), Measuring and Dating Business Cycles of the Economy of Iran, \emph{Journal of Monetary & Banking Research}, 7(20), Summer 2014, pp. 161-194. (in Persian)}