\name{MBRI.Iran.Dating}
\alias{MBRI.Iran.Dating}
\docType{data}
\title{
Dating of Business Cycles of Iran by MBRI}
\description{
This is the official Dating of Business Cycles of Iran by MBRI. This is not exactly what you get using \code{\link[BCDating:BBQ]{BBQ}} on \code{\link[BCDating:Iran.non.Oil.GDP.Cycle]{Iran.non.Oil.GDP.Cycle}} as there are some changes to that based on other economic facts. See reference paper for details.
}
\usage{
data(MBRI.Iran.Dating)
}
\format{BCDating Object}
\source{Einian, M. and M. Barakchian (2012)}
\references{Einian, M. and M. Barakchian (2012), Measuring and Dating Business Cycles of the Economy of Iran, \emph{Pool va Eqtesad}, forthcoming (in Persian)}
\examples{
data(MBRI.Iran.Dating)
plot(MBRI.Iran.Dating)













}
\keyword{datasets}