\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,BCDating-method}
\title{Showing a BCDating object}
\description{Methods for function \code{show}}
\section{Methods}{
\describe{

\item{\code{signature(object = "BCDating")}}{
Shows the dates of peaks and troughs of the BCDating.
}
}}
\author{
Majid Einian,\email{m.einian@mbri.ac.ir}, \cr\href{http://www.mbri.ac.ir}{Monetary and Banking Research Institute}, \href{http://www.cbi.ir}{Central Bank of Islamic Republic of Iran}

Franck Arnaud
}
\keyword{methods}
\examples{
library(BCDating)
data("MBRI.Iran.Dating")
MBRI.Iran.Dating
}
