\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,BCDating-method}
\title{Summerizing a BCDating Object}
\description{Methods for function \code{summary}}
\section{Methods}{
\describe{

\item{\code{signature(object = "BCDating")}}{
Lists the start and end dates of recessions and expansions in a BCDating, their duration, amplitude ,... Also the average duration of expansions and recessions are printed.
}
}}
\keyword{methods}
\author{
Majid Einian,\email{m.einian@mbri.ac.ir}, \cr\href{http://www.mbri.ac.ir}{Monetary and Banking Research Institute}, \href{http://www.cbi.ir}{Central Bank of Islamic Republic of Iran}

Franck Arnaud
}
