\name{Iran.non.Oil.GDP.Quarterly.Growth}
\docType{data}
\alias{Iran.non.Oil.GDP.Quarterly.Growth}
\title{
Quartely Grwoth of non-Oil GDP of Iran.}
\description{
Quartely Grwoth of non-Oil GDP of Iran. (after x12)
}
\usage{
Iran.non.Oil.GDP.Quartely.Grwoth
}
\format{ts Quarterly Time Series}
\source{Central Bank of Islamic Republic of Iran. Further calculations by Majid Einian}
\references{Einian, M. and M. Barakchian (2012), Measuring and Dating Business Cycles of the Economy of Iran, \emph{Pool va Eqtesad}, forthcoming (in Persian)}